Create Rest Parameter
Rest parameter allows us to have a variable no. of parameters, like earlier, we needed to specify the no. of parameter in the function definition, and there could be only that many parameter, no more no less(could be if function definition doesn't use the absent parameter in computation. ). But with this we need not worry about the restriction of no. of parameters.